+++ ciphers/src/subtle/hash_functions/mod.rs
@@ -1,2 +1,4 @@
+pub mod blake2b;
+pub mod incorrect_hmac_blake2b;
 pub mod keyed_shake256;
 mod infer_keyed_hash;
